requirements.txt
hikari
python-dotenv
hikari-lightbulb
hikari-miru
humanize